Rename rl_common directory to common
