+++ include/agents/DiscretizationAgent.hh
@@ -1,8 +1,8 @@
 #ifndef _DISCAGENT_HH_
 #define _DISCAGENT_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 
 #include <vector>
