+++ include/agents/Dyna.hh
@@ -5,8 +5,8 @@
 #ifndef _DYNA_HH_
 #define _DYNA_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 
 #include <map>
 #include <set>
