+++ include/agents/ModelBasedAgent.hh
@@ -6,8 +6,8 @@
 #ifndef _MODELBASED_HH_
 #define _MODELBASED_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 
 #include <set>
 #include <vector>
