+++ include/environments/FuelRooms.hh
@@ -9,8 +9,8 @@
 #define _FUELROOMS_H_
 
 #include <set>
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 #include "gridworld.hh"
