+++ include/environments/LightWorld.hh
@@ -8,8 +8,8 @@
 #define _LIGHTWORLD_H_
 
 #include <set>
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 
 
 class LightWorld: public Environment {
