+++ include/environments/RobotCarVel.hh
@@ -12,8 +12,8 @@
 #define _ROBOTCAR_H_
 
 #include <set>
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 
 /** This class defines a domain that is a simulation of velocity control for the Austin Robot 
     Technology autonomous vehicle. */
