+++ include/environments/energyrooms.hh
@@ -2,8 +2,8 @@
 #define _ENERGYROOMS_H_
 
 #include <set>
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 #include "gridworld.hh"
 
 /*
