+++ include/environments/taxi.hh
@@ -9,8 +9,8 @@
 #define _TAXI_H_
 
 #include <set>
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 #include "gridworld.hh"
 
 /** This class defines the Taxi domain. */
