+++ include/models/C45Tree.hh
@@ -9,8 +9,8 @@
 #ifndef _C45TREE_HH_
 #define _C45TREE_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 #include <vector>
 #include <set>
 #include <map>
