+++ include/models/ExplorationModel.hh
@@ -9,8 +9,8 @@
 #ifndef _EXPLOREMODEL_HH_
 #define _EXPLOREMODEL_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 #include <vector>
 #include <map>
 #include <set>
