+++ include/models/M5Tree.hh
@@ -8,8 +8,8 @@
 #ifndef _M5TREE_HH_
 #define _M5TREE_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 #include <vector>
 #include <set>
 #include <map>
