+++ include/models/MultipleModels.hh
@@ -1,8 +1,8 @@
 #ifndef _MULTIPLEMODELS_HH_
 #define _MULTIPLEMODELS_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 #include <vector>
