+++ include/models/RMaxModel.hh
@@ -6,8 +6,8 @@
 #ifndef _RMAXMODEL_HH_
 #define _RMAXMODEL_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 #include <vector>
 #include <map>
 #include <set>
