+++ include/models/SepPlanExplore.hh
@@ -8,8 +8,8 @@
 
 #include "models/Stump.hh"
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 #include <vector>
 #include <set>
 #include <map>
