+++ include/models/Stump.hh
@@ -1,8 +1,8 @@
 #ifndef _STUMP_HH_
 #define _STUMP_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 #include <vector>
 #include <set>
 #include <map>
