+++ include/planners/ETUCT.hh
@@ -10,8 +10,8 @@
 #ifndef _ETUCT_HH_
 #define _ETUCT_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 
 #include "models/FactoredModel.hh"
