+++ include/planners/PO_ETUCT.hh
@@ -10,8 +10,8 @@
 #ifndef _PO_ETUCT_HH_
 #define _PO_ETUCT_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 
 #include "models/FactoredModel.hh"
