+++ include/planners/PrioritizedSweeping.hh
@@ -1,8 +1,8 @@
 #ifndef _PRIORITIZEDSWEEPING_HH_
 #define _PRIORITIZEDSWEEPING_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 
 #include <set>
 #include <vector>
