+++ include/planners/ValueIteration.hh
@@ -6,8 +6,8 @@
 #ifndef _VALUEITERATION_HH_
 #define _VALUEITERATION_HH_
 
-#include <rl_common/Random.h>
-#include <rl_common/core.hh>
+#include <common/Random.h>
+#include <common/core.hh>
 
 #include <set>
 #include <vector>
